refactor(core): split llm.ts into focused modules
Split the 1801-line LLM service file into 7 focused modules:
- schemas.ts: Zod schemas and exported types
- prompts.ts: Shared prompt constants
- object-repair.ts: JSON repair strategies for malformed LLM output
- plan-generation.ts: Two-phase and single-phase plan generation
- code-generation.ts: Code generation, analysis, and error recovery
- llm-service.ts: Core LLMService class (model creation, text/object generation)
- factory.ts: createLLMService factory function

The original llm.ts is preserved as a thin re-export shim so all existing
imports continue to work unchanged.

Closes #30
