+++ packages/core/src/llm/index.ts
@@ -1,0 +1,3 @@
+export { LLMService, normalizeProviderBaseURL } from './llm-service.js';
+export { createLLMService } from './factory.js';
+export type { GeneratedPlan, GeneratedCode, GeneratedPatch } from './schemas.js';
